Move isChain near connection
