+++ src/custom/components/Header/NetworkSelector/NetworkSelectorMod.tsx
@@ -2,7 +2,7 @@ import { ArrowDownCircle } from 'react-feather'
 import styled from 'styled-components/macro'
 import { Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
-import { getWeb3ReactConnection } from '@cow/modules/wallet/web3-react/connection'
+import { getWeb3ReactConnection, isChainAllowed } from '@cow/modules/wallet/web3-react/connection'
 import { getChainInfo } from 'constants/chainInfo'
 import { CHAIN_IDS_TO_NAMES, SupportedChainId } from 'constants/chains'
 import useParsedQueryString from 'hooks/useParsedQueryString'
@@ -17,7 +17,7 @@ import { updateConnectionError } from 'state/connection/reducer'
 import { useAppDispatch } from 'state/hooks'
 import { ExternalLink, MEDIA_WIDTHS } from 'theme'
 import { replaceURLParam } from 'utils/routes'
-import { isChainAllowed, switchChain } from 'utils/switchChain'
+import { switchChain } from '@cow/modules/wallet/web3-react/hooks/switchChain'
 // import { isMobile } from 'utils/userAgent'
 
 // Mod imports
